working chains add
